% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_magic_number.R
\name{add_magic_number}
\alias{add_magic_number}
\title{Expand magic numbers data: Add row containing seed and text}
\usage{
add_magic_number(.seed, .text, .data = MagicNumbers)
}
\arguments{
\item{.seed}{seed that generates .text when used to sample()}

\item{.text}{text generated by sample, used after set.seed(seed)}

\item{.data}{data frame to be expanded; defaults to MagicNumbers}
}
\value{
data frame that contains the new row of data
}
\description{
Expand magic numbers data: Add row containing seed and text
}
\examples{
add_magic_number(.text = "love", .seed = 249725)
MagicNumbers <- add_magic_number(.text = "love", .seed = 249725)
}
