% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_seed.R
\name{check_seed_number}
\alias{check_seed_number}
\title{Check if a random seed generates a specific number (one or more digits)}
\usage{
check_seed_number(seed, size, number)
}
\arguments{
\item{seed}{random seed, used in set.seed(seed)}

\item{size}{number of digits, used in sample(..., size = size)}

\item{number}{number to check}
}
\value{
TRUE if seed generates the number, FALSE otherwise
}
\description{
Check if a random seed generates a specific number (one or more digits)
}
\examples{
check_seed_number(1, 4, "1000")  # FALSE
}
