% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_seed.R
\name{check_seed_phrase}
\alias{check_seed_phrase}
\title{Check if a random seed generates a specific phrase (letters and blanks)}
\usage{
check_seed_phrase(seed, size, phrase)
}
\arguments{
\item{seed}{random seed, used in set.seed(seed)}

\item{size}{number of characters of word, used in sample(..., size = size)}

\item{phrase}{phrase to check: may contain lower-case letters and blanks}
}
\value{

}
\description{
Check if a random seed generates a specific phrase (letters and blanks)
}
