% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_seed.R
\name{check_seed_word}
\alias{check_seed_word}
\title{Check if a random seed generates a specific word (only letters, no blanks)}
\usage{
check_seed_word(seed, size, word)
}
\arguments{
\item{seed}{random seed, used in set.seed(seed)}

\item{size}{number of characters of word, used in sample(..., size = size)}

\item{word}{word to check: may contain only lower-case letters, no blanks}
}
\value{
TRUE if seed generates the word, FALSE otherwise
}
\description{
Check if a random seed generates a specific word (only letters, no blanks)
}
\examples{
internal function, called from seek_seeds
}
