% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_seedwords_par.R
\name{iterate_seedwords_par}
\alias{iterate_seedwords_par}
\title{Iterate over a vector of seeds to find a magic number for your word of choice. Runs in parallel using clusterApply()}
\usage{
iterate_seedwords_par(.seeds, .word, cl)
}
\arguments{
\item{.seeds}{vector of seeds to iterate over, e. g. 0:1e5}

\item{.word}{word of choice that you want to find a magic number for}

\item{cl}{cluster object for clusterApply(). cl <- makeCluster(parallelly::availableCores(omit = 1))}
}
\value{
Currently no useful return value. In case of success, calculation stops, the magic number (seed) is displayed in the console.
}
\description{
Iterate over a vector of seeds to find a magic number for your word of choice. Runs in parallel using clusterApply()
}
\examples{
# Set up parallelization
library(parallel)
cl <- makeCluster(parallelly::availableCores(omit = 1))
iterate_seedwords_par(0:1e3, "I", cl)
}
