% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MagicNumbers.R
\docType{data}
\name{MagicNumbers}
\alias{MagicNumbers}
\title{Magic Numbers = seeds that generate specific texts in sample(...)}
\format{
A data frame with 2 variables and, hopefully, a steadily growing number of rows:
\describe{
\item{text}{Text to be generated by sample()}
\item{seed}{The magic number to be used in set.seed(seed) before sample()}
}
}
\usage{
MagicNumbers
}
\description{
Dataset containing magic numbers that can be used in set.seed() to generate text accordingly in sample(...).
NA in seed means: no seed found in valid range of seeds from 0 to 2147483647.
}
\keyword{datasets}
