% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seek_seeds.R
\name{seek_seeds_number}
\alias{seek_seeds_number}
\title{Iterate over a vector of seeds to find a magic number for your number of choice.}
\usage{
seek_seeds_number(.seeds, .number, .progress = FALSE)
}
\arguments{
\item{.seeds}{vector of seeds to iterate over, e. g. 0:1e5}

\item{.number}{number of choice that you want to find a magic number for}

\item{.progress}{Show progress bar (TRUE) or not (FALSE)}
}
\value{
Currently no useful return value. In case of success, calculation stops, the magic number (seed) is displayed in the console.
}
\description{
Iterate over a vector of seeds to find a magic number for your number of choice.
}
\examples{
seek_seeds_number(0:1e3, "2023")
}
