% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_old_seek_seeds.R, R/seek_seeds.R
\name{seek_seeds_word}
\alias{seek_seeds_word}
\title{Iterate over a vector of seeds to find a magic number for your word of choice.}
\usage{
seek_seeds_word(.seeds, .word, .eval = "seq", cl = NULL, .progress = FALSE)

seek_seeds_word(.seeds, .word, .eval = "seq", cl = NULL, .progress = FALSE)
}
\arguments{
\item{.seeds}{vector of seeds to iterate over, e. g. 0:1e5}

\item{.word}{word of choice that you want to find a magic number for}

\item{.eval}{type of evaluation: "seq" uses parallel::lapply(), "par" uses clusterApply, "future" uses future.apply::future_lapply}

\item{cl}{cluster object for clusterApply(). cl <- makeCluster(parallelly::availableCores(omit = 1))}

\item{.progress}{Show progress bar (TRUE) or not (FALSE); ignored if .eval = "par"}
}
\value{
Currently no useful return value. In case of success, calculation stops, the magic number (seed) is displayed in the console.

Currently no useful return value. In case of success, calculation stops, the magic number (seed) is displayed in the console.
}
\description{
Iterate over a vector of seeds to find a magic number for your word of choice.

Iterate over a vector of seeds to find a magic number for your word of choice.
}
\examples{
seek_seeds_word(0:1e3, "ab")
seek_seeds_word(50000:70000, "xmas", .eval = "par", cl)
seek_seeds_word(50000:70000, "xmas", .eval = "future", cl)
seek_seeds_word(0:1e3, "ab")
seek_seeds_word(50000:70000, "xmas", .eval = "par", cl)
seek_seeds_word(50000:70000, "xmas", .eval = "future", cl)
}
