% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_seed.R
\name{test_seed}
\alias{test_seed}
\title{Test if a given seed returns a sensible text when used to sample()}
\usage{
test_seed(seed, nchar, choices = letters)
}
\arguments{
\item{seed}{to be used in set.seed(seed)}

\item{nchar}{length of expected text = number of random one character samples}

\item{choices}{what to choose from; defaults to letters; can be c(letters, " ") to allow for phrases}
}
\value{

}
\description{
Test if a given seed returns a sensible text when used to sample()
}
\examples{
test_seed(19421956, 5)
test_seed(64927, 4)
}
